requirements.txt
openai
groq
pytrends
requests
beautifulsoup4
streamlit
python-dotenv
pillow
reportlab
gtts
moviepy
watchdog

